\name{MMCoxCL-package}
\alias{MMCoxCL-package}
\alias{MMCoxCL}
\docType{package}
\title{
\packageTitle{MMCoxCL}
}
\description{
\packageDescription{MMCoxCL}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{MMCoxCL}
\packageIndices{MMCoxCL}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{MMCoxCL}

Maintainer: \packageMaintainer{MMCoxCL}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file ~~
~~ KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
